MH-13190, Refactor JpaGroupRoleProvider JaxRs REST concerns into a separate class           Mitigate opencast-userdirectory getService cycle (load race condition with JaxRsServiceTracker.addingService)           Give external-api control of its API response
