+++ modules/admin-ui/src/main/java/org/opencastproject/adminui/endpoint/GroupsEndpoint.java
@@ -51,8 +51,10 @@ import org.opencastproject.matterhorn.search.SearchResult;
 import org.opencastproject.matterhorn.search.SearchResultItem;
 import org.opencastproject.matterhorn.search.SortCriterion;
 import org.opencastproject.security.api.SecurityService;
+import org.opencastproject.security.api.UnauthorizedException;
 import org.opencastproject.security.api.User;
 import org.opencastproject.security.api.UserDirectoryService;
+import org.opencastproject.userdirectory.ConflictException;
 import org.opencastproject.util.NotFoundException;
 import org.opencastproject.util.RestUtil;
 import org.opencastproject.util.data.Option;
@@ -250,7 +252,16 @@ public class GroupsEndpoint {
       @RestResponse(responseCode = SC_NOT_FOUND, description = "Group not found."),
       @RestResponse(responseCode = SC_INTERNAL_SERVER_ERROR, description = "An internal server error occured.")})
   public Response removeGroup(@PathParam("id") String groupId) throws NotFoundException {
-    return indexService.removeGroup(groupId);
+    try {
+      indexService.removeGroup(groupId);
+      return Response.noContent().build();
+    } catch (NotFoundException e) {
+      return Response.status(SC_NOT_FOUND).build();
+    } catch (UnauthorizedException e) {
+      return Response.status(SC_FORBIDDEN).build();
+    } catch (Exception e) {
+      throw new WebApplicationException(e);
+    }
   }
 
   @POST
@@ -271,7 +282,17 @@ public class GroupsEndpoint {
       @RestResponse(responseCode = SC_CONFLICT, description = "An group with this name already exists.") })
   public Response createGroup(@FormParam("name") String name, @FormParam("description") String description,
           @FormParam("roles") String roles, @FormParam("users") String users) {
-    return indexService.createGroup(name, description, roles, users);
+    try {
+      indexService.createGroup(name, description, roles, users);
+    } catch (IllegalArgumentException e) {
+      logger.warn(e.getMessage());
+      return Response.status(Status.BAD_REQUEST).build();
+    } catch (UnauthorizedException e) {
+      return Response.status(SC_FORBIDDEN).build();
+    } catch (ConflictException e) {
+      return Response.status(SC_CONFLICT).build();
+    }
+    return Response.status(Status.CREATED).build();
   }
 
   @PUT
@@ -295,7 +316,15 @@ public class GroupsEndpoint {
   public Response updateGroup(@PathParam("id") String groupId, @FormParam("name") String name,
           @FormParam("description") String description, @FormParam("roles") String roles,
           @FormParam("users") String users) throws NotFoundException {
-    return indexService.updateGroup(groupId, name, description, roles, users);
+    try {
+      indexService.updateGroup(groupId, name, description, roles, users);
+    } catch (IllegalArgumentException e) {
+      logger.warn(e.getMessage());
+      return Response.status(Status.BAD_REQUEST).build();
+    } catch (UnauthorizedException ex) {
+      return Response.status(SC_FORBIDDEN).build();
+    }
+    return Response.ok().build();
   }
 
   @GET
